Put everything in the same struct
