--- src/gui.c
@@ -1,3 +1,0 @@
-#import <stdio.h>
-#import <stdlib.h>
-#import <string.h>
